+++ src/emc/sai/saicanon.cc
@@ -900,7 +900,7 @@ void SET_MOTION_CONTROL_MODE(double tolerance) { }
 
 int GET_EXTERNAL_DIGITAL_INPUT(int index, int def) { return def; }
 double GET_EXTERNAL_ANALOG_INPUT(int index, double def) { return def; }
-int WAIT(int index, int input_type, int wait_type, int timeout) { return 0; }
+int WAIT(int index, int input_type, int wait_type, double timeout) { return 0; }
 
 /* Returns the system feed rate */
 double GET_EXTERNAL_FEED_RATE()
